copyright methods extracted in their own copyrightUtils.js
